streamlit_dashboard/requirements.txt
streamlit>=1.27
pandas>=1.5
snowflake-connector-python>=3.0
python-dotenv>=1.0
plotly>=5.18
